+++ R/install.R
@@ -13,6 +13,13 @@
 #' separate package installation backends in the future.
 #'
 #'
+#' @section Remotes Syntax:
+#'
+#' `renv` supports a subset of the `remotes` syntax used for package installation,
+#' as described in <https://remotes.r-lib.org/articles/dependencies.html>. See
+#' the examples below for more details.
+#'
+#'
 #' @section Package Configuration:
 #'
 #' Many \R packages have a `configure` script that needs to be run to prepare
@@ -56,6 +63,13 @@
 #'
 #' # install 'digest' from GitHub (latest dev. version)
 #' renv::install("eddelbuettel/digest")
+#'
+#' # install a package from GitHub, using specific commit
+#' renv::install("eddelbuettel/digest@@df55b00bff33e945246eff2586717452e635032f")
+#'
+#' # install a package from Bioconductor
+#' # (note: requires the BiocManager package)
+#' renv::install("bioc::Biobase")
 #'
 #' # install a package from local sources
 #' renv::install("~/path/to/package")