+++ tests/testthat/test-remotes.R
@@ -22,7 +22,7 @@ test_that("we can parse a variety of remotes", {
   expect_equal(record$Package, "skeleton")
   expect_equal(record$Version, "1.0.1")
   expect_equal(record$RemoteRef, "master")
-  expect_equal(record$RemoteSha, "958296dbbbf7f1d82f7f5dd1b121c7558604809f")
+  expect_equal(record$RemoteSha, "e4aafb92b86ba7eba3b7036d9d96fdfb6c32761a")
 
   # by commit
   record <- renv_remotes_resolve("kevinushey/skeleton@209c4e48e505e545ad7ab915904d983b5ab83b93")